# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocialPredict API
  version: 0.1.0
  description: >
    HTTP contract for the SocialPredict backend service. This spec is generated
    from the Go handlers and DTOs under backend/server and backend/handlers.
    Routes wrapped by the shared security middleware can return a shared
    transport-level `429 text/plain` response before reaching handler-specific
    JSON envelope logic when the rate limit is exceeded. Per-operation `429`
    responses document that middleware behavior where it applies.
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
...
  /v0/changepassword:
    post:
      operationId: changePassword
      tags: [Users]
      summary: Change account password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordEnvelopeResponse'
        '400':
          description: Invalid request body or password requirements not met.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
                validationFailed:
                  value:
                    ok: false
                    reason: VALIDATION_FAILED
        '401':
          description: Authentication is required or the current password is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                authenticationRequired:
                  value:
                    ok: false
                    reason: AUTHENTICATION_REQUIRED
                invalidCredentials:
                  value:
                    ok: false
                    reason: INVALID_CREDENTIALS
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error while changing password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
  /v0/usercredit/{username}:
    get:
      operationId: getUserCredit
      tags: [Users]
      summary: Get user credit
      description: Returns the user's available credit limit, treating missing users as max debt.
      parameters:
        - in: path
          name: username
          required: true
          description: Username to evaluate for credit availability.
          schema:
            type: string
      responses:
        '200':
          description: Credit calculated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreditResponse'
        '400':
          description: Username missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Failed to calculate user credit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
  /v0/portfolio/{username}:
    get:
      operationId: getUserPortfolio
      tags: [Users]
      summary: Get user portfolio
      description: Returns the markets and share counts for a user's portfolio. Unknown users currently resolve to the same empty portfolio shape as users with no recorded bets.
      parameters:
        - in: path
          name: username
          required: true
          description: Username whose portfolio should be returned.
          schema:
            type: string
      responses:
        '200':
          description: Portfolio returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
        '400':
          description: Username missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while fetching portfolio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    MiddlewareRateLimitExceeded:
      description: Request rejected by shared rate-limiting middleware before handler-specific logic ran.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainTextErrorResponse'
  schemas:
...
    FailureReason:
      type: string
      description: |
        Canonical public failure reasons for the swept API routes covered by API-002 through API-008:
        - `METHOD_NOT_ALLOWED`: the route rejected the HTTP method before domain work ran.
        - `INVALID_REQUEST`: request JSON, path, or query input was malformed.
        - `VALIDATION_FAILED`: caller-supplied data failed validation or sanitization.
        - `AUTHENTICATION_REQUIRED`: a protected route was called without valid authentication.
        - `AUTHORIZATION_DENIED`: the caller is authenticated but lacks required privileges.
        - `PASSWORD_CHANGE_REQUIRED`: the caller must rotate their password before continuing.
        - `INVALID_CREDENTIALS`: submitted credentials were rejected during login or password checks.
        - `NOT_FOUND`: the requested user, market, content, or position target does not exist.
        - `MARKET_CLOSED`: market state prevents the requested market action.
        - `INSUFFICIENT_BALANCE`: accounting rules reject the requested action.
        - `INVALID_OUTCOME`: the supplied YES/NO outcome is not valid for the operation.
        - `NO_POSITION`: no eligible position exists for the requested sale.
        - `INSUFFICIENT_SHARES`: the sale exceeds the caller's held shares.
        - `DUST_CAP_EXCEEDED`: the requested sale would exceed the configured dust cap.
        - `INTERNAL_ERROR`: sanitized fallback for unexpected server-side failures.
      enum:
        - METHOD_NOT_ALLOWED
        - INVALID_REQUEST
        - VALIDATION_FAILED
        - AUTHENTICATION_REQUIRED
        - AUTHORIZATION_DENIED
        - PASSWORD_CHANGE_REQUIRED
        - INVALID_CREDENTIALS
        - NOT_FOUND
        - MARKET_CLOSED
        - INSUFFICIENT_BALANCE
        - INVALID_OUTCOME
        - NO_POSITION
        - INSUFFICIENT_SHARES
        - DUST_CAP_EXCEEDED
        - INTERNAL_ERROR
      example: VALIDATION_FAILED
    ReasonResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        reason:
          $ref: '#/components/schemas/FailureReason'
      required: [ok, reason]
    PlainTextErrorResponse:
      type: string
      description: Plain text error emitted by shared middleware before request-specific handlers run.
...
    ChangePasswordEnvelopeResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        result:
          $ref: '#/components/schemas/ChangePasswordResult'
      required: [ok, result]
    ChangePasswordResult:
      type: object
      properties:
        message:
          type: string
          example: Password changed successfully
      required: [message]
...
    UserCreditResponse:
      type: object
      properties:
        credit:
          type: integer
          format: int64
    PortfolioItem:
      type: object
      properties:
        marketId:
          type: integer
        questionTitle:
          type: string
        yesSharesOwned:
          type: integer
          format: int64
        noSharesOwned:
          type: integer
          format: int64
        lastBetPlaced:
          type: string
          format: date-time
    PortfolioResponse:
      type: object
      properties:
        portfolioItems:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
        totalSharesOwned:
          type: integer
          format: int64
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...